fix: decoupled nodeRed utility
